requirements.txt
aiogram>=3.4.1,<4.0
redis>=5.0.0,<6.0
aioredis>=2.0.0,<3.0
cryptography==41.0.7
python-dotenv==1.0.0
structlog==23.3.0
pydantic>=2.7.0,<3.0
gunicorn>=21.0.0